Changed (double) to (float) so that we don't need special code on NVIDIA platforms.
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@3272 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ tests/kernels/test_cl_random.cl
@@ -54,7 +54,7 @@ cl_random_state(uint4 state)
 
 static inline float cl_random_prob(uint4 state)
 {
-   return (float) ((double) state.s0 / (double) 4294967296.0);
+   return (float) ((float) state.s0 / (float) 4294967296.0);
 }
